chore: 빌드 시, document 객체에 접근하지 못하는 원인 파악 테스트 (#101)
